requirements.txt
asyncpg==0.27.0
psutil==5.9.5
pytest==7.4.0
pytest-asyncio==0.21.0
pyyaml==6.0
grpcio==1.71.0rc2
grpcio-tools
protobuf
tensorflow-gpu==2.11.0
keras==2.11.0
tensorboard==2.11.0
torch==2.6.0
torchaudio==2.6.0
torchvision==0.21.0
textual==0.10.0
nats-py==2.2.0
transformers==4.31.0
datasets==2.14.4
beautifulsoup4==4.12.2
requests
numpy==1.25.0
scikit-learn==1.3.0
cffi==1.15.1
